[update] 增加控制最大尺寸参数。

+++ documentskewcorrection-core/src/main/java/io/github/alexmofer/documentskewcorrection/core/DocumentSkewDetectorCanny.java
@@ -43,7 +43,7 @@ public final class DocumentSkewDetectorCanny extends DocumentSkewDetector {
         private Bitmap mImage;
         private boolean mRecycleImage;
 
-        public Builder(int maxSize) {
+        public Builder(float maxSize) {
             mMaxSize = maxSize;
         }
